Shuffled some things around for easier re-use.
